// src/images/images.d.ts
declare const images: {
    bannerSrc: any;
    logoSrc: any;
    siteLogo: any;
    noNotification: any;
    artistBg: any;
    brandBg: any;
    systemBg: any;
    defaultAvatar: any;
    placeholder: any;
    add: any;
    faceBrowser: any;
    faceDocument: any;
    faceHeart: any;
    faceLock: any;
    faceStop: any;
    chromeIcon: any;
    safariIcon: any;
    edgeIcon: any;
    external: any;
    allLinksGrid: any;
    gearIcon: any;
    defaultIcon: any;
    remove: any;
    defaultIconCircle: any;
    spinner: any;
};
export default images;
//# sourceMappingURL=images.d.ts.map