// src/shared/types/genericTypes.ts
import { ArticleState } from './articleTypes';
import { BookmarkState } from './bookmarkTypes';
import { ContentStateModel } from './contentTypes';
import { DepartmentsState } from './departmentTypes';
import { EmployeeState } from './employeeTypes';
import { EventState } from './eventTypes';
import { LinkState } from './linkTypes';
import { ManualNotificationListState } from './manualNotificationTypes';
import { NotificationState } from './notificationTypes';
import { OfficesPeopleState } from './officesPeopleType';
import { OfficeState } from './officeTypes';
import { PageState } from './pageTypes';
import { PeopleState } from './peopleTypes';
import { ProfileState } from './profileTypes';
import { SearchResultState } from './searchResultTypes';
import { TodoState } from './todoTypes';
import { UserProfileState } from './userProfileTypes';
import { UsersState } from './userTypes';
import { CommonState } from './commonTypes';
import { AppreciationState } from './appreciationTypes';
import { LabelState } from './labelTypes';
import { SystemsApplicationsState } from './systemsApplicationsTypes';

export interface AppState {
  appreciationReducers: AppreciationState;
  articleReducers: ArticleState;
  bookmarkReducers: BookmarkState;
  contentReducers: ContentStateModel;
  departmentReducers: DepartmentsState;
  labelReducers: LabelState;
  departmentsReducers: DepartmentsState;
  employeeReducers: EmployeeState;
  eventReducers: EventState;
  linkReducers: LinkState;
  manualNotificationReducers: ManualNotificationListState;
  notificationReducers: NotificationState;
  officeReducers: OfficeState;
  officesPeopleReducers: OfficesPeopleState;
  pageReducers: PageState;
  peopleReducers: PeopleState;
  profileReducers: ProfileState;
  searchResultReducers: SearchResultState;
  todoReducers: TodoState;
  userProfilePictureReducers: UserProfileState;
  userProfileReducers: UserProfileState;
  usersReducers: UsersState;
  commonReducers: CommonState;
  systemsApplicationsReducers: SystemsApplicationsState;
}